+++ src/calendar.html
@@ -60,9 +60,9 @@
     <my-sidebar></my-sidebar>
     
     <!-- Scripts -->
-    <script type="module" src="scripts/components/calendar-view.js"></script>
-    <script type="module" src="scripts/components/task-pop-up.js"></script>
-    <script src="scripts/components/sidebar.js"></script>
+    <script type="module" src="scripts/components/calendar/view.js"></script>
+    <script type="module" src="scripts/components/task/pop-up.js"></script>
+    <script src="scripts/components/layout/sidebar.js"></script>
 
 </body>
 </html>
