+++ src/index.html
@@ -27,11 +27,10 @@
 </head>
 
 <body>
-    <script type="module" src="scripts/components/upcoming-tasks.js"></script>
-    <script src="scripts/components/sidebar.js"></script>
-    <script src="scripts/components/modal-journal.js"></script>
-    <script type="module" src="scripts/components/task-component.js"></script>
-    <script type="module" src="scripts/components/task-pie-component.js"></script>
+    <script type="module" src="scripts/components/task/upcoming-tasks.js"></script>
+    <script src="scripts/components/layout/sidebar.js"></script>
+    <script type="module" src="scripts/components/task/view.js"></script>
+    <script type="module" src="scripts/components/task/pie-chart.js"></script>
 
     <div class="grid">
     <h1>Task<br>Overview</h1>
