+++ src/journal.html
@@ -26,13 +26,13 @@
 </head>
 
 <body>
-    <script src="scripts/components/sidebar.js"></script>
-    <script src="scripts/components/journal_tree.js" type="module"></script>
-    <script src="scripts/components/journal-editor.js" type="module"></script>
+    <script src="scripts/components/layout/sidebar.js"></script>
+    <script src="scripts/components/journal/tree.js" type="module"></script>
+    <script src="scripts/components/journal/editor.js" type="module"></script>
 
     <!-- Include the JavaScript file defining the Web Component -->
-    <script src="scripts/components/modal-journal.js" type="module"></script>
-    <script src="scripts/components/journal-main.js" type="module"></script>
+    <script src="scripts/components/journal/modal.js" type="module"></script>
+    <script src="scripts/components/journal/view.js" type="module"></script>
 
     <!-- Use the custom Web Component here -->
     <!-- <modal-journal></modal-journal> -->
