+++ src/scripts/components/journal/view.js
