+++ src/scripts/components/layout/sidebar.js
