+++ src/tasks.html
@@ -33,7 +33,7 @@
             <main></main>
             <!--Add a Column Box--> 
         </div>
-        <script type="module" src="scripts/components/task.js"></script>
-        <script src="scripts/components/sidebar.js"></script>
+        <script type="module" src="scripts/components/task/kanban.js"></script>
+        <script src="scripts/components/layout/sidebar.js"></script>
     </body>
 </html>
